requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
lagom==2.7.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# GraphQL
strawberry-graphql[fastapi]==0.219.2
graphql-core==3.2.3

# AI & ML
langchain==0.1.6
langgraph==0.0.25
langchain-openai==0.0.5
langchain-google-genai==0.0.11
openai==1.12.0
google-generativeai==0.3.2

# RAG & Vector Stores
faiss-cpu==1.7.4
weaviate-client==4.4.2
chromadb==0.4.22
sentence-transformers==2.3.1

# Databases
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
redis[hiredis]==5.0.1

# HTTP & WebSockets
httpx==0.26.0
aiohttp==3.9.3
websockets==12.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0

# Security
python-dotenv==1.0.1
cryptography==42.0.2

# Testing
coverage==7.4.1
unittest-xml-reporting==3.2.0
faker==22.6.0
factory-boy==3.3.0

# Code Quality
ruff==0.2.1
mypy==1.8.0
pre-commit==3.6.0

# Documentation
pyyaml==6.0.1
